🎨 Co-locate all admin-related code in /admin
refs #4172

- move all the admin related code from controllers, routes and helpers into a single location
- refactor the way assets are registered, to keep everything working for now
- duplicate & fix the checkSSL middleware to use req.originalURL
- remove the decide-is-admin middleware
- move the redirect-to-setup middleware into the admin
- TODO: remove the ugly hack for loading assets
- TODO: remove the duplicated checkSSL call
- TODO: rethink the structure, this should probably be an internal app
