+++ core/test/unit/middleware/check-ssl_spec.js
@@ -25,7 +25,7 @@ describe('checkSSL', function () {
     });
 
     it('should not require SSL (frontend)', function (done) {
-        req.url = '/';
+        req.originalUrl = '/';
         checkSSL(req, res, next);
         next.called.should.be.true();
         next.calledWith().should.be.true();
@@ -33,7 +33,7 @@ describe('checkSSL', function () {
     });
 
     it('should require SSL (frontend)', function (done) {
-        req.url = '/';
+        req.originalUrl = '/';
         req.secure = true;
         checkSSL(req, res, next);
         next.called.should.be.true();
@@ -42,7 +42,7 @@ describe('checkSSL', function () {
     });
 
     it('should not require SSL (admin)', function (done) {
-        req.url = '/ghost';
+        req.originalUrl = '/ghost';
         res.isAdmin = true;
         checkSSL(req, res, next);
         next.called.should.be.true();
@@ -51,7 +51,7 @@ describe('checkSSL', function () {
     });
 
     it('should not redirect with SSL (admin)', function (done) {
-        req.url = '/ghost';
+        req.originalUrl = '/ghost';
         res.isAdmin = true;
         res.secure = true;
 
@@ -62,7 +62,7 @@ describe('checkSSL', function () {
     });
 
     it('should not redirect with force admin SSL (admin)', function (done) {
-        req.url = '/ghost';
+        req.originalUrl = '/ghost';
         res.isAdmin = true;
         req.secure = true;
         configUtils.set({
@@ -76,7 +76,7 @@ describe('checkSSL', function () {
     });
 
     it('should redirect with force admin SSL (admin)', function (done) {
-        req.url = '/ghost/';
+        req.originalUrl = '/ghost/';
         res.isAdmin = true;
         res.redirect = {};
         req.secure = false;
@@ -97,7 +97,7 @@ describe('checkSSL', function () {
     });
 
     it('should redirect to subdirectory with force admin SSL (admin)', function (done) {
-        req.url = '/blog/ghost/';
+        req.originalUrl = '/blog/ghost/';
         res.isAdmin = true;
         res.redirect = {};
         req.secure = false;
@@ -118,7 +118,7 @@ describe('checkSSL', function () {
     });
 
     it('should redirect and keep query with force admin SSL (admin)', function (done) {
-        req.url = '/ghost/';
+        req.originalUrl = '/ghost/';
         req.query = {
             test: 'true'
         };
@@ -142,7 +142,7 @@ describe('checkSSL', function () {
     });
 
     it('should redirect with with config.url being SSL (frontend)', function (done) {
-        req.url = '/';
+        req.originalUrl = '/';
         req.secure = false;
         res.redirect = {};
         configUtils.set({
@@ -162,7 +162,7 @@ describe('checkSSL', function () {
     });
 
     it('should redirect to urlSSL (admin)', function (done) {
-        req.url = '/ghost/';
+        req.originalUrl = '/ghost/';
         res.isAdmin = true;
         res.redirect = {};
         req.secure = false;
@@ -183,7 +183,7 @@ describe('checkSSL', function () {
     });
 
     it('should not redirect if redirect:false (admin)', function (done) {
-        req.url = '/ghost/';
+        req.originalUrl = '/ghost/';
         res.isAdmin = true;
         res.sendStatus = {};
         req.secure = false;
